config.py
import os
from dotenv import load_dotenv

# Cargar variables de entorno desde el archivo .env
basedir = os.path.abspath(os.path.dirname(__file__))
load_dotenv(os.path.join(basedir, '.env'), override=True)

class Config:
    # Clave secreta para proteger formularios y sesiones
    SECRET_KEY = os.environ.get('SECRET_KEY') or 'dev-key-por-defecto'
    
    # Configuración de la base de datos (PostgreSQL en Supabase vía pg8000)
    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URL')
    if SQLALCHEMY_DATABASE_URI:
        if SQLALCHEMY_DATABASE_URI.startswith("postgres://"):
            SQLALCHEMY_DATABASE_URI = SQLALCHEMY_DATABASE_URI.replace("postgres://", "postgresql+pg8000://", 1)
        elif SQLALCHEMY_DATABASE_URI.startswith("postgresql://"):
            SQLALCHEMY_DATABASE_URI = SQLALCHEMY_DATABASE_URI.replace("postgresql://", "postgresql+pg8000://", 1)
            
    import ssl
    ssl_ctx = ssl.create_default_context()
    ssl_ctx.check_hostname = False
    ssl_ctx.verify_mode = ssl.CERT_NONE

    SQLALCHEMY_ENGINE_OPTIONS = {
        "connect_args": {
            "ssl_context": ssl_ctx
        }
    }
    
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    
    # Configuración de Supabase
    SUPABASE_URL = os.environ.get("SUPABASE_URL")
    SUPABASE_KEY = os.environ.get("SUPABASE_KEY")
    
    # Configuración original (para compatibilidad de migraciones si la usas local)
    UPLOAD_FOLDER = os.path.join('app', 'static', 'uploads', 'products')
    MAX_CONTENT_LENGTH = 16 * 1024 * 1024  # Límite de 16 MB para subidas